Move auth logic to `auth.py` and don't do auth if app is public
1. Move `auth` logic to `auth.py` removes some of the
deployment-specific lines in `app.py`. This makes `app.py` closer
resemble the users own `app.py` that they might have created while
experimenting with Dash
2. Before, we were still registering the app with streambed even if the
app was public. Now, we just skip that step. This means that the user
does not need to supply their api key or username or understand what
their `PLOTLY_DOMAIN` is when they are getting started. Hopefully this
streamlines the users first deployment process.
This is at the expense of not having the apps appear in organize.
@cldougl and I talked about this, and we felt like:
1 - Most users seem to be only working from the dash app manager
2 - Having public apps organize doesn’t provide much value. They can
change the permissions (to public) but they should maybe just be doing
that inside `config.py`. There is still some value if the apps are
private, in case they want to add collaborators.
There is risk that this behaviour will become confusing: only private
apps will appear in organize, not public apps. I think this potential
confusion is worth the streamlined first-deployment experience.